Add first implimentation of connect via
This change will allow octahe to connect to host via some set of jump
hosts assuming all hosts are using the same ssh key. While this is not
currently using libssh2 (more research is needed on that) it does enable
the advertised capability, retains the speed we need to be successful
and is a clean solution.

Fixes #14

Signed-off-by: Kevin Carter <kecarter@redhat.com>