Add initial support for the FROM verb
This change adds initial (basic) support for the FROM verb which is
referenced in issue #5. This change will allow us to asynchronous
query various registries for container images and return image history
when discovered. Initial support has been limited to quay.io, however
I will work on dockerhub next.

Once we have a method for adding various public registries and querying
them, I will begin on a process to allow for Octahe to use authentication
credentials, and private registries; these are problems for sometime in
the after now.

Fixes #5

Signed-off-by: Kevin Carter <kecarter@redhat.com>