when using via ensure we generate a one time use ssh config
Instead of passing in crazy params on the CLI and hoping every user
has a compatible version of the ssh client installed, we're going to
use a temp ssh config which will generate the required config.

This is less than ideal, but it works well and is compatible with most
ssh clients going back for the last few years.

Signed-off-by: Kevin Carter <kecarter@redhat.com>